Tighten UFFD pager review boundaries
