+++ lib/hypervisor/hypervisor.go
@@ -115,13 +115,32 @@ type VMStarter interface {
 	// - Cloud Hypervisor: starts process, calls Restore API
 	// - QEMU: would start with -incoming or -loadvm flags (not yet implemented)
 	// Returns the process ID and a Hypervisor client. The VM is in paused state after restore.
-	RestoreVM(ctx context.Context, p *paths.Paths, version string, socketPath string, snapshotPath string) (pid int, hv Hypervisor, err error)
+	RestoreVM(ctx context.Context, p *paths.Paths, version string, socketPath string, snapshotPath string, opts RestoreOptions) (pid int, hv Hypervisor, err error)
 
 	// PrepareFork allows hypervisors to prepare forked instance state.
 	// For snapshot-based forks, implementations can rewrite snapshot config with
 	// fork identity (paths, vsock, network). Hypervisors that don't support fork
 	// should return ErrNotSupported.
 	PrepareFork(ctx context.Context, req ForkPrepareRequest) (ForkPrepareResult, error)
+}
+
+type SnapshotMemoryBackend string
+
+const (
+	SnapshotMemoryBackendFile SnapshotMemoryBackend = "file"
+	SnapshotMemoryBackendUFFD SnapshotMemoryBackend = "uffd"
+)
+
+type SnapshotMemoryOverlay struct {
+	GuestMemoryOffset int64
+	Path              string
+}
+
+type RestoreOptions struct {
+	SnapshotMemoryBackend   SnapshotMemoryBackend
+	SnapshotMemoryCacheKey  string
+	SnapshotMemoryOverlays  []SnapshotMemoryOverlay
+	SnapshotMemorySessionID string
 }
 
 // ForkNetworkConfig contains network identity fields for fork preparation.
