+++ lib/hypervisor/tracing.go
@@ -296,7 +296,7 @@ func (s *tracingVMStarter) StartVM(ctx context.Context, p *paths.Paths, version 
 	return pid, hv, err
 }
 
-func (s *tracingVMStarter) RestoreVM(ctx context.Context, p *paths.Paths, version string, socketPath string, snapshotPath string) (pid int, hv Hypervisor, err error) {
+func (s *tracingVMStarter) RestoreVM(ctx context.Context, p *paths.Paths, version string, socketPath string, snapshotPath string, opts RestoreOptions) (pid int, hv Hypervisor, err error) {
 	ctx, span := startTraceSpan(ctx, s.tracer, "hypervisor.restore_vm",
 		attribute.String("hypervisor", string(s.hvType)),
 		attribute.String("operation", "restore_vm"),
@@ -307,7 +307,7 @@ func (s *tracingVMStarter) RestoreVM(ctx context.Context, p *paths.Paths, versio
 		}
 		finishTraceSpan(span, err)
 	}()
-	pid, hv, err = s.next.RestoreVM(ctx, p, version, socketPath, snapshotPath)
+	pid, hv, err = s.next.RestoreVM(ctx, p, version, socketPath, snapshotPath, opts)
 	return pid, hv, err
 }
